Add column PID + fix ColumnType.LONG

+++ src/me/corriekay/pokegoutil/data/enums/ColumnType.java
@@ -24,7 +24,7 @@ public enum ColumnType {
         CellRenderers.NUMBER
     ),
     LONG(
-        Integer.class,
+        Long.class,
         Comparators.LONG,
         CellRenderers.NUMBER
     ),
